[C++] Split library into mcapreader.hpp and mcapwriter.hpp, introduce MCAP_IMPLEMENTATION macro
Fixes ODR violations by conditionally including .inl files when MCAP_IMPLEMENTATION is defined
