+++ cpp/bench/run.cpp
@@ -1,4 +1,5 @@
-#include <mcap/mcap.hpp>
+#define MCAP_IMPLEMENTATION
+#include <mcap/mcapwriter.hpp>
 
 #include <benchmark/benchmark.h>
