+++ cpp/test/unit-tests.cpp
@@ -1,3 +1,4 @@
+#define MCAP_IMPLEMENTATION
 #include <mcap/mcap.hpp>
 
 #include <array>
